requirements.txt
py-cord==2.6.0
discord==2.3.2
pymongo==4.8.0
certifi==2024.8.30
python-dotenv==1.0.1